requirements.txt
Django==2.0.7
djangorestframework==3.8.2
pycodestyle==2.4.0
Pygments==2.2.0
pytz==2018.5
